Change ReadResFile/WriteResFile interface to take size_t args; removes 64KB size limitation on key-value files; bug #1112
git-svn-id: https://sc2.svn.sourceforge.net/svnroot/sc2/trunk/sc2@3558 8092fc87-c524-0410-9efc-e669fe64eaf9

+++ src/libs/resource/propfile.c
@@ -98,7 +98,7 @@ PropFile_from_string (char *d, PROPERTY_HANDLER handler, const char *prefix)
 void
 PropFile_from_file (uio_Stream *f, PROPERTY_HANDLER handler, const char *prefix)
 {
-	long flen;
+	size_t flen;
 	char *data;
 
 	flen = LengthResFile (f);
@@ -108,6 +108,8 @@ PropFile_from_file (uio_Stream *f, PROPERTY_HANDLER handler, const char *prefix)
 		return;
 	}
 
+	// We may end up with less bytes than we asked for due to the
+	// DOS->Unix newline conversion
 	flen = ReadResFile (data, 1, flen, f);
 	data[flen] = '\0';
